libft

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: lraffin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/11 00:32:58 by lraffin           #+#    #+#             */
-/*   Updated: 2021/10/13 04:23:13 by lraffin          ###   ########.fr       */
+/*   Updated: 2021/10/13 04:56:51 by lraffin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s && s[i])
